2of3: sign recovery key at subaccount creation
At creation of 2of3 subaccounts, sign the recovery key with the
login key and pass it to the server.

Generation of backup mnemonic and derivation of the xpub are moved
to the "resolver" also for the case without an hardware device.

Since the message signed includes the subaccount number, such value
is always obtained in the "resolver", which caused some changes in
the internal interface.